Switched expression widgets that should represent double precision floats to ExpressionType.DOUBLE

+++ menus/shared_widgets/color_selector.py
@@ -58,7 +58,7 @@ class ColorSelectorWidget(SquareWidget):
         self.position_layout = QHBoxLayout()
         self.master_layout.addLayout(self.position_layout)
 
-        self.position_le = ExpressionWidget(position, ExpressionType.FLOAT, parent=self, clamps=(0.0, 100.0))
+        self.position_le = ExpressionWidget(position, ExpressionType.DOUBLE, parent=self, clamps=(0.0, 100.0))
         self.position_layout.addWidget(self.position_le, 1)
         self.position_percent_label = QLabel("%")
         self.position_percent_label.setAlignment(Qt.AlignLeft)
